refactor: JDBCのequipmentクラスのパッケージ移動
refs #002
